Python: Split out debpic.py into multiple files
