+++ debian/debpic.install
@@ -7,5 +7,5 @@ debpic/Documentation/ /usr/share/doc/debpic/
 debpic/debpic-completion.bash => /usr/share/bash-completion/completions/debpic
 debpic/debpic.conf.template => /etc/debpic/debpic.conf.template
 debpic/Dockerfile => usr/share/debpic/Dockerfile
-debpic/debpic.py => usr/bin/debpic
 debpic/hooks/ /etc/debpic/
+debpic/python/*.py usr/lib/python3/dist-packages/debpic
