+++ debian/debpic.links
@@ -1,0 +1,1 @@
+usr/lib/python3/dist-packages/debpic/debpic.py /usr/bin/debpic
